This is an attempt to solve #3 and #16, allowing the result of `validate` to have multiple errors in the same `nested` array so that users can solve several problems in their input data at once.

At the moment this is a proof of concept, there's still quite a bit of work to do and decisions to be made, including whether to go forward with this at all.

Most types don't need to be able to report multiple errors so they haven't changed. Some types we could consider adding this to but I'm reluctant about:

- TArray: if you get one item wrong, several others are probably wrong for the same reason.
- TTuple and TParamList: if you forget an item before the end or get the order wrong, this will cause several errors with the same fix.
- TUnion: it might be nice to show why a few candidates failed instead of just one best guess, but this is questionable.

So for now it's just TIface and TIntersection. TIface is a good example to show how this works.

First, this line is unchanged:

```ts
if (typeof value !== "object" || value === null) { return ctx.fail(null, "is not an object", 0); }
```

If it's not an object, there's nothing else to report, so we call `ctx.fail` directly and then return early. But from here on, `fail` must no longer be called directly on `ctx`, at least until this checker returns. Failures will instead be stored in *forks*. These could be renamed to branches, children, subcontexts, etc. They should be used as follows:

- Get a forked context with `ctx.fork()`
- Use the fork rather than `ctx` for (potential) failures, whether it's in a deeper checker as in `propCheckers[i](v, fork)` or directly as in `fork.fail(name, null, 1)`.
- After using the fork to check one 'thing' (e.g. a property or a base class), write:

```ts
          if (!ctx.completeFork()) {
            return false;
          }
```

- The main context might make use of the information that you're done with the latest fork.
- It then returns true if there's 'room' for more errors and you should continue checking, or false if you should just return now.
- Once you're done checking, `return !ctx.failed()` to check whether any failures were gathered along the way without triggering an early return.

The NoopContext still does very little, only storing a boolean and returning itself from `fork()`. But the extra operations have some inevitable overhead. I haven't done proper calculations yet but at a glance the benchmark (great thing to have!) seems to run something like 10-15% slower.

The DetailContext keeps forks that have failures in them. The way it does this right now is not very clever and will probably be optimised. The number of failed forks it keeps is limited (right now to 3, but that will be made configurable) so that users don't get flooded with errors. Once it reaches that limit, `ctx.completeFork()` returns false to indicate you should exit. But since you can now have a whole tree of contexts, the total number of forks and errors is unlimited.

Some things that still need addressing:

- Formatting a nice hierarchical multiline VError message when necessary
- A proper API - I have many thoughts on this topic
- The issue I mentioned in https://github.com/gristlabs/ts-interface-checker/pull/4#issuecomment-808807995

What are your thoughts so far?