@alexmojaki 2021-03-30
I think the API needs some redesigning. The names `test`, `check`, and `validate` are easy to confuse, there's a `strict` copy of each one, and it's hard to add more options without creating a combinatorial explosion of methods. In particular I want to add options for https://github.com/gristlabs/ts-interface-checker/pull/4#issuecomment-808807995 and https://github.com/alexmojaki/ts-interface-checker/pull/1, and there may be more in the future such as an option to deal with #46 or #38.

I propose encouraging one primary central method where the user configures everything, something like this:

```ts
export class Checker {

    public validate(value: any, options: CheckOptions): IErrorDetail | null {
       ...
    }

    public check(value: any, options: CheckOptions): void {
        this.validate(value, {...options, onError: OnError.throw});
    }

    public test(value: any, options: CheckOptions): boolean {
        return this.validate(
            value,
            {...options, onError: OnError.return, numErrors: 0}
        ) == null;
    }
    
    ...
}

interface CheckOptions {
    extras: Extras
    onError: OnError

    /**
     * Number of errors that can be nested directly
     * under another error.
     *
     * Internally, the number of forks a DetailContext can hold.
     *
     * 0: Don't record anything.
     *   - Used for test() to just return a boolean
     *   - Could be used in `check()` if someone wants to raise an error
     *      but doesn't want a message.
     *   - Creates a NoopContext
     * 1: Old default behaviour for check/validate
     * > 1: Can nest several errors
     *   - May lead to a multiline error message when using OnError.throw
     *   - Requires checkRoot to be true
     */
    numErrors: number

    /**
     * If true, wraps the top level error in another layer about the
     * root type being checked. For example, this would change
     * the error message from:
     *
     *     value.extras is missing
     *
     * to:
     *
     *     value is not a CheckOptions; value.extras is missing
     *
     * Similarly the returned IErrorDetail will have an extra top layer.
     *
     * This is required if numErrors > 1 so that multiple errors can be
     * nested under the root IErrorDetail.
     *
     * Internally this wraps the type being checked in a TName first.
     */
    checkRoot: boolean
}

/** What to do when extraneous values are found */
enum Extras {
    /** Allow extra values */
    ignore,

    /** Formerly strict mode: do what onError says */
    error,

    /** Remove extra values in place */
    delete,
}

/** What to do when an error is encountered */
enum OnError {
    /** Throw a VError */
    throw,

    /** Return an object describing the errors */
    return,
}
```

Questions:

- Do you like this idea?
- What should the default values for options be? Keep behaviour the same? Default `numErrors` to 3? Require specifying certain options?
- Deprecate some old methods? I suggest deprecating the `strict` methods, but leaving `check` and `test` as is for the clean return types.
- How to deal with conflicting options, e.g. `numErrors: 1, checkRoot: true`, or `check(value, {onError: OnError.throw})`? Just overwrite them, or throw an error?