@alexmojaki 2021-04-05
> Agreed that an `options` object is more flexible.
> 
> I like adding an option for `extraProps` to make "strict" versions redundant, and to add the option to delete extra properties. So perhaps all methods can take `options` with `{extraProps?: 'error' | 'delete'}`, by default omitted.

So if I understand correctly:

1. `options` will only have one key `extraProps` and none of the others mentioned, so its purpose (as opposed to an `extraProps` argument) will be to make adding more options in the future easier.
2. Both `options` itself and `extraProps` are optional.

By the way, I called it `extras` because in general it also deals with extra items in tuples and parameter lists and I didn't know if those are considered 'properties', but I'm happy with `extraProps` since interface properties are what people usually have in mind anyway.

> For the distinction between validate / check / test, I think using separate methods is a bit better than a single method with different return types -- in particular for type-checking, and explaining what it does. As long as these methods exist, I don't see much purpose in emulating them by passing options to some unified method.
> 
> Last but not least, for the level of details in errors, I think more informative errors are just better. The only reason against making them the _only_ behavior is if there is existing code that relies on the previous format. I am thinking, if we bump the major version and explain the difference, it would be fine to just upgrade the errors to contain more details, and avoid burdening the user with more options.

Right, the idea of a central method with lots of options was that the user would have one clear place to look to discover all their needs, but if we just reduce the number of options there's less need for that and I like your thinking.

So shall the number of errors per object just be fixed at 3?