@alexmojaki 2021-04-10
I've done a detailed benchmark analysis. Seems like it's now 16% slower.

Disclaimer: I'm not a statistician. 

<details>
<summary>Click for full analysis</summary>

```python
import re
import statistics

output = """
$ npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && git checkout master && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench && npm run bench

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,191,607 ops/sec ±0.20% (97 runs sampled)
protobuf verify x 8,823,535 ops/sec ±0.29% (96 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,189,434 ops/sec ±0.17% (98 runs sampled)
protobuf verify x 8,702,189 ops/sec ±0.21% (98 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,138,205 ops/sec ±0.19% (98 runs sampled)
protobuf verify x 8,205,776 ops/sec ±2.26% (93 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,147,411 ops/sec ±0.51% (97 runs sampled)
protobuf verify x 8,533,095 ops/sec ±0.52% (95 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,047,723 ops/sec ±0.18% (95 runs sampled)
protobuf verify x 7,943,914 ops/sec ±2.28% (90 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,080,907 ops/sec ±0.21% (96 runs sampled)
protobuf verify x 8,196,217 ops/sec ±0.18% (96 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 996,332 ops/sec ±0.20% (92 runs sampled)
protobuf verify x 7,708,242 ops/sec ±0.31% (91 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,056,713 ops/sec ±0.16% (93 runs sampled)
protobuf verify x 7,847,602 ops/sec ±0.43% (96 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,101,959 ops/sec ±0.17% (95 runs sampled)
protobuf verify x 8,127,860 ops/sec ±0.26% (96 runs sampled)
Switched to branch 'master'
Your branch is up to date with 'origin/master'.

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,216,527 ops/sec ±0.24% (97 runs sampled)
protobuf verify x 7,780,665 ops/sec ±0.18% (94 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,297,354 ops/sec ±0.22% (95 runs sampled)
protobuf verify x 8,057,131 ops/sec ±0.17% (97 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,174,733 ops/sec ±0.29% (95 runs sampled)
protobuf verify x 7,451,642 ops/sec ±0.25% (96 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,246,502 ops/sec ±0.44% (88 runs sampled)
protobuf verify x 7,660,006 ops/sec ±0.28% (93 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,236,564 ops/sec ±0.46% (93 runs sampled)
protobuf verify x 7,919,159 ops/sec ±0.20% (94 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,269,222 ops/sec ±0.17% (92 runs sampled)
protobuf verify x 7,855,862 ops/sec ±0.29% (97 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,260,112 ops/sec ±0.17% (96 runs sampled)
protobuf verify x 7,858,326 ops/sec ±0.18% (94 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,195,676 ops/sec ±0.22% (97 runs sampled)
protobuf verify x 7,504,485 ops/sec ±0.20% (94 runs sampled)

> ts-interface-checker@0.2.1 bench
> tsc && node test/bench/bench.js

ts-interface-checker x 1,178,706 ops/sec ±0.23% (97 runs sampled)
protobuf verify x 7,514,983 ops/sec ±0.90% (93 runs sampled)
"""

numbers = [int(n.replace(",", "")) for n in re.findall(r"[\d,]{7,}", output)]

half = len(numbers) // 2

# 'before' was printed second, after git checkout master
before = numbers[half:]
after = numbers[:half]
assert len(before) == len(after)
assert len(before) % 2 == 0


def average_ratio(nums):
    ratios = []
    for i in range(0, len(nums), 2):
        tsi, proto = nums[i:i + 2]
        ratio = tsi / proto
        ratios.append(ratio)
    print(ratios)
    # After:
    # [0.13504870780248507, 0.13668216123552362, 0.13870778339550094, 0.13446598215536099, 0.1318900229785972, 0.13187876797307832, 0.1292554125830507, 0.13465425489213137, 0.13557799962105646]
    # Before:
    # [0.15635257397664595, 0.16101935043627813, 0.1576475359390588, 0.16272859316298185, 0.15614839909136816, 0.1615636832724404, 0.1603537445506842, 0.1593281884099975, 0.15684746059971127]
    # Higher ratios are better
    return statistics.geometric_mean(ratios)


print(average_ratio(after) / average_ratio(before))
# 0.8436143737304612
# i.e. 84.3% as fast as before
```
</details>